requirements.txt
python-can
python-socketio
aiohttp
